requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2023.5.7
distlib==0.3.6
dj-database-url==2.1.0
Django==4.2.1
django-cors-headers==4.3.0
django-database-url==1.0.3
django-filter==23.3
djangorestframework==3.14.0
filelock==3.12.2
flake8==6.0.0
gunicorn==21.2.0
mccabe==0.7.0
numpy==1.24.3
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
pipenv==2023.6.18
platformdirs==3.5.3
plotly==5.14.1
psycopg==3.1.12
psycopg2-binary==2.9.9
pycodestyle==2.10.0
pyflakes==3.0.1
PyMySQL==1.1.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tenacity==8.2.2
typing_extensions==4.8.0
tzdata==2023.3
virtualenv==20.23.0
virtualenv-clone==0.5.7
